+++ scripts/run_tests.sh
@@ -9,9 +9,9 @@ set -e
 # Run code quality checks before tests; a failure here
 # will prevent the tests from running.
 source_dirs="/app/relation_engine_server /app/client_src /app/spec"
-flake8 --max-complexity 20 "$source_dirs"
-mypy --ignore-missing-imports "$source_dirs"
-bandit -r "$source_dirs"
+flake8 --max-complexity 20 $source_dirs
+mypy --ignore-missing-imports $source_dirs
+bandit -r $source_dirs
 
 # start server, using the specs in /spec/repo
 sh /app/scripts/start_server.sh &
@@ -21,8 +21,12 @@ coverage erase
 python -m spec.validate
 
 # run importer/, relation_engine_server/, and spec/ tests
-coverage run --branch -m unittest discover -v importers relation_engine_server spec
+coverage run --branch --parallel-mode -m unittest discover -v importers
+coverage run --branch --parallel-mode -m unittest discover -v relation_engine_server
+coverage run --branch --parallel-mode -m unittest discover -v spec
+
 
 # RE client tests
 PYTHONPATH=client_src python -m unittest discover client_src/test
+coverage combine
 coverage html --omit=*/test_*